Extract MetricConnectionCallback into network
To reuse it in the acra-server later.
